backend/requirements.txt
Django>=5.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
spacy>=3.7.0
PyPDF2>=3.0.0
python-docx>=1.1.0
